+++ ibc-testkit/src/relayer/mod.rs
@@ -1,2 +1,7 @@
 pub mod context;
 pub mod error;
+pub mod utils;
+
+// `ibc::apps::transfer::handler::send_transfer` requires `serde`
+#[cfg(feature = "serde")]
+pub mod integration;
